# analytics-engine/openapi/analytics-engine-api.yaml
openapi: 3.0.3
info:
  title: Analytics Engine API
  description: |
    Production-ready Analytics Engine API for AI video analytics, human tracking,
    vehicle analytics, industrial safety, and real-time detection processing.
    
    **Core Features:**
    - Real-time frame processing with ONNX inference
    - 14 analytics modules (human, vehicle, industrial, retail, banking, etc.)
    - Person/vehicle tracking with Re-ID
    - Heat map generation and crowd analytics
    - Industrial safety monitoring
    - SNMP device integration
    - Camera health monitoring
    
    **Performance:**
    - <100ms person detection latency
    - 1-2 FPS frame processing
    - 30+ concurrent camera streams (CPU)
    - 50-80+ streams (GPU)
    
    **Authentication:**
    All endpoints require `x-analytics-source-key` or `x-analytics-engine-key` header.
  version: 2.0.0
  contact:
    name: Analytics Engine Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8092
    description: Development server
  - url: https://analytics.sentinelgrid.com
    description: Production server
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/analytics/industrial/zones/{zoneId}:
    get:
      tags: [Industrial Analytics]
      summary: Get zone details
      operationId: getIndustrialZone
      parameters:
        - name: zoneId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Zone details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/IndustrialZone'

    patch:
      tags: [Industrial Analytics]
      summary: Update zone
      operationId: updateIndustrialZone
      parameters:
        - name: zoneId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndustrialZoneUpdate'
      responses:
        '200':
          description: Zone updated

    delete:
      tags: [Industrial Analytics]
      summary: Delete zone
      operationId: deleteIndustrialZone
      parameters:
        - name: zoneId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Zone deleted
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-analytics-source-key
      description: Analytics engine API key
  schemas:
...
    IndustrialZone:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        camera_id:
          type: string
        zone_name:
          type: string
        zone_type:
          type: string
          enum: [restricted, equipment-only, pedestrian-only, hazard, monitoring]
        polygon_points:
          type: array
          items:
            type: object
            properties:
              x:
                type: number
              y:
                type: number
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    IndustrialZoneUpdate:
      type: object
      properties:
        zoneName:
          type: string
        enabled:
          type: boolean
        polygonPoints:
          type: array
          items:
            type: object
...